Refactored files and fixed coverage issue
