+++ src/modules/common/hooks/index.js
@@ -3,4 +3,3 @@ export * from './useCustomInfiniteQuery';
 export * from './useOutsideClick';
 export { default as useSort } from './useSort';
 export { default as useFilter } from './useFilter';
-export { default as useCheckLegacyAccount } from './useCheckLegacyAccount';
