+++ src/modules/legacy/hooks/queries/index.js
@@ -1,1 +1,2 @@
-export * from './useLegacy'
+export * from './useLegacy';
+export { default as useCheckLegacyAccount } from './useCheckLegacyAccount';
